% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ROC_functions.r
\name{AUC.FUN}
\alias{AUC.FUN}
\title{AUC function}
\usage{
AUC.FUN(data)
}
\arguments{
\item{data}{}
}
\value{
AUC
}
\description{
AUC function
}
\examples{
AUC.FUN(cbind(rbinom(10,size=2,prob=.5),rnorm(10)))
}

