% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autofeature_functions.R
\name{emr_select}
\alias{emr_select}
\title{Automated feature selection}
\usage{
emr_select(dat.X, dat.S, b0 = 100)
}
\arguments{
\item{dat.X}{data.frame or matrix of covariates}

\item{dat.S}{data.frame or matrix of surrogate markers}

\item{b0}{number of perturbation resampling draws}
}
\value{
list
}
\description{
Automated feature selection
}

