% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{Est.ALASSO.GLM.Approx}
\alias{Est.ALASSO.GLM.Approx}
\title{GLM with approximation}
\usage{
Est.ALASSO.GLM.Approx(data, Wi = NULL, rtn = "EST", nopen.ind = NULL,
  BIC.factor = 0.1, offset = NULL)
}
\arguments{
\item{data}{1st column y; remaining x}

\item{Wi}{weights for resampling if interested in obtaining standard errors}

\item{nopen.ind}{indexes which subset of x should not be penalized}

\item{offset}{}
}
\value{
bhat
}
\description{
GLM with approximation
}

