% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{Est.ALASSO.GLM}
\alias{Est.ALASSO.GLM}
\title{Logistic regression or logistic adaptive LASSO estimator.
Initial estimator obtained via ridge}
\usage{
Est.ALASSO.GLM(data, Wi = NULL, rtn = "EST", nopen.ind = NULL,
  regularize = yes.regularize, BIC.factor = 0.1, offset = NULL,
  fam0 = "binomial")
}
\arguments{
\item{data}{1st column y; remaining x}

\item{Wi}{weights for resampling if interested in obtaining standard errors}

\item{rtn}{}

\item{nopen.ind}{}

\item{regularize}{binary; if FALSE estimates GLM, if TRUE estimates adaptive LASSO GLM}

\item{BIC.factor}{}

\item{offset}{}

\item{fam0}{}
}
\value{
beta
}
\description{
Logistic regression or logistic adaptive LASSO estimator.
Initial estimator obtained via ridge
}

