% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missclassification_functions.R
\name{Par.FUN}
\alias{Par.FUN}
\title{Title}
\usage{
Par.FUN(dat)
}
\arguments{
\item{dat}{data matrix with "Y", variables beginning with "S", variables
beginning with "G"}
}
\value{
beta, alpha1, alpha0
}
\description{
Title
}
\examples{
Par.FUN()
}

