% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofeature_functions.R
\name{ProbD.S}
\alias{ProbD.S}
\title{Calculate Probability Y=1 based on mclust output}
\usage{
ProbD.S(Si, par)
}
\arguments{
\item{Si}{data.frame or matrix of surrogate markers}

\item{par}{parameter output from mclust function}
}
\value{
numeric value of length nrow(Si) estimating probability Y=1
}
\description{
Calculate Probability Y=1 based on mclust output
}

