% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autofeature_functions.R
\name{ProbD.SX}
\alias{ProbD.SX}
\title{Predict D=1 via mclust on S and X}
\usage{
ProbD.SX(dat.Xt, dat.St, dat.Xv, dat.Sv, betahat, mclust.modelNames = "EEI")
}
\arguments{
\item{dat.Xt}{}

\item{dat.St}{}

\item{dat.Xv}{}

\item{dat.Sv}{}

\item{betahat}{}

\item{mclust.modelNames}{}

\item{mclustfit.par}{}
}
\value{
list of predictive probabilities based on X, S, and S+X for training and validation sets
}
\description{
Predict D=1 via mclust on S and X
}
\examples{
#update with an example dataset
#kernfit <- kern_varselect(dat.S=dat.S,dat.X=dat.X,b0=0,mclust.modelNames="EEI")
#predall <- ProbD.SX(dat.Xt,dat.St,dat.Xv,dat.Sv,betahat=kernfit$bhat[-1])
#AUC.FUN(cbind(dat.Yv,predall$pi.SXv))
}

