% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SIM.FUN.MultiG}
\alias{SIM.FUN.MultiG}
\title{Simulate multiG}
\usage{
SIM.FUN.MultiG(nn, b.G, b.S, rho.S)
}
\arguments{
\item{nn}{}

\item{b.G}{}

\item{b.S}{}

\item{rho.S}{}
}
\value{
dat
}
\description{
Simulate multiG
}

