% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SIM.FUN}
\alias{SIM.FUN}
\title{Simulate data}
\usage{
SIM.FUN(nn, b.G, b.S, rho.S)
}
\arguments{
\item{nn}{}

\item{b.G}{}

\item{b.S}{}

\item{rho.S}{}
}
\value{
dat
}
\description{
Simulate data
}
\examples{
b0.G = rep(c(0.3,0.5,0.3,0.2,0),c(1,2,3,2,50)); b0.G = rbind(-1,b0.G); p = ncol(b0.G)
b0.S = c(1.5,1.5,2); rho0.S = 0.3; A0 = length(b0.S);
mydata0 = SIM.FUN(NN,b.G=b0.G,b.S=b0.S,rho.S=rho0.S);
}

