% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{VTM}
\alias{VTM}
\title{Vector to matrix}
\usage{
VTM(vc, dm)
}
\arguments{
\item{vc}{vector}

\item{dm}{number of rows}
}
\value{
matrix
}
\description{
Vector to matrix
}
\examples{
VTM(1:10,3)
}

