% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autofeature_functions.R
\name{kern_varselect}
\alias{kern_varselect}
\title{Kernel function to perform surrogate clustering,
regularized regression estimation, and
perturbation resampling}
\usage{
kern_varselect(dat.S, dat.X, b0)
}
\arguments{
\item{dat.S}{}

\item{dat.X}{}

\item{b0}{}
}
\value{
list
}
\description{
Kernel function to perform surrogate clustering,
regularized regression estimation, and
perturbation resampling
}

