% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SIM.FUN}
\alias{SIM.FUN}
\title{Simulate data}
\usage{
SIM.FUN(nn, b.G, b.S, rho.S, rho.G, ng.cts, Gtype = "binomial",
  Stype = "cts")
}
\arguments{
\item{nn}{}

\item{b.G}{}

\item{b.S}{}

\item{rho.S}{}

\item{rho.G}{}

\item{ng.cts}{}

\item{Gtype}{}

\item{Stype}{}
}
\value{
dat
}
\description{
Simulate data
}
\examples{
b0.G = rep(c(0.3,0.5,0.3,0.2,0),c(1,2,3,2,50)); b0.G = rbind(-1,b0.G); p = ncol(b0.G)
b0.S = c(1.5,1.5,2)
mydata0 = SIM.FUN(100,b.G=b0.G,b.S=b0.S,rho.S=0.3,ng.cts=3,rho.G=0.2);
}

